feat: update cli to read from tokenami.config.js
