+++ package.json
@@ -4,10 +4,9 @@
   "version": "1.0.0",
   "description": "",
   "scripts": {
-    "prepare": "pnpm build-fast",
+    "prepare": "pnpm build",
     "dev": "pnpm --filter=./examples/remix dev",
-    "build": "pnpm -r --parallel --filter=./packages/** build",
-    "build-fast": "pnpm -r --parallel --filter=./packages/** build-fast"
+    "build": "pnpm -r --parallel --filter=./packages/** build"
   },
   "devDependencies": {
     "prettier": "^2.8.8",
