+++ packages/cli/bin.js
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 
-require('./dist/index.js');
+import './dist/index.js';
