+++ packages/cli/package.json
@@ -2,9 +2,9 @@
   "name": "@tokenami/cli",
   "version": "1.0.0",
   "description": "",
+  "type": "module",
   "main": "dist/index.js",
   "module": "dist/index.mjs",
-  "sideEffects": false,
   "bin": {
     "tokenami": "bin.js"
   },
@@ -14,13 +14,15 @@
     "bin.js"
   ],
   "scripts": {
-    "build": "tsup src --format=esm,cjs --dts --no-splitting --shims",
-    "build-fast": "tsup src --format=esm,cjs --no-dts --no-splitting --shims",
-    "dev": "pnpm build-fast --watch src"
+    "build": "tsup src --format=esm --dts --no-splitting --shims",
+    "dev": "pnpm build --watch src"
   },
   "devDependencies": {
     "@types/node": "^20.3.1",
     "ts-node": "^10.9.1",
     "typescript": "^5.1.3"
+  },
+  "dependencies": {
+    "cac": "^6.7.14"
   }
 }
