+++ packages/cli/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "include": ["src"],
+  "exclude": ["node_modules"],
+  "compilerOptions": {
+    "target": "esnext",
+    "module": "esnext",
+    "moduleResolution": "node"
+  }
+}
